move models' elasticsearch support to traits
