+++ app/Models/Beatmapset.php
@@ -37,7 +37,7 @@ use Illuminate\Database\QueryException;
 
 class Beatmapset extends Model
 {
-    use EsIndexable, SoftDeletes;
+    use Elasticsearch\BeatmapsetTrait, SoftDeletes;
 
     protected $_storage = null;
     protected $table = 'osu_beatmapsets';
@@ -88,6 +88,63 @@ class Beatmapset extends Model
     const RANKED_PER_DAY = 8;
     const MINIMUM_DAYS_FOR_RANKING = 7;
     const BUNDLED_IDS = [3756, 163112, 140662, 151878, 190390, 123593, 241526, 299224];
+
+    /*
+    |--------------------------------------------------------------------------
+    | Elasticsearch mappings; can't put in a Trait.
+    |--------------------------------------------------------------------------
+    */
+    const ES_MAPPINGS_BEATMAPS = [
+        'approved' => ['type' => 'long'],
+        'beatmap_id' => ['type' => 'long'],
+        'countNormal' => ['type' => 'long'],
+        'countSlider' => ['type' => 'long'],
+        'countSpinner' => ['type' => 'long'],
+        'countTotal' => ['type' => 'long'],
+        'diff_approach' => ['type' => 'double'],
+        'diff_drain' => ['type' => 'double'],
+        'diff_overall' => ['type' => 'double'],
+        'diff_size' => ['type' => 'double'],
+        'difficultyrating' => ['type' => 'double'],
+        'hit_length' => ['type' => 'long'],
+        'passcount' => ['type' => 'long'],
+        'playcount' => ['type' => 'long'],
+        'playmode' => ['type' => 'long'],
+        'total_length' => ['type' => 'long'],
+        'version' => ['type' => 'string'],
+    ];
+
+    const ES_MAPPINGS_BEATMAPSETS = [
+        'approved' => ['type' => 'long'],
+        'approved_date' => ['type' => 'date'],
+        'artist' => ['type' => 'string'],
+        'artist_unicode' => ['type' => 'string'],
+        'bpm' => ['type' => 'double'],
+        'creator' => ['type' => 'string'],
+        'difficulty_names' => ['type' => 'string'],
+        'download_disabled' => ['type' => 'boolean'],
+        'epilepsy' => ['type' => 'boolean'],
+        'favourite_count' => ['type' => 'long'],
+        'filename' => ['type' => 'string'],
+        'filesize' => ['type' => 'long'],
+        'filesize_novideo' => ['type' => 'long'],
+        'genre_id' => ['type' => 'long'],
+        'language_id' => ['type' => 'long'],
+        'last_update' => ['type' => 'date'],
+        'offset' => ['type' => 'long'],
+        'play_count' => ['type' => 'long'],
+        'rating' => ['type' => 'double'],
+        'source' => ['type' => 'string'],
+        'star_priority' => ['type' => 'long'],
+        'storyboard' => ['type' => 'boolean'],
+        'submit_date' => ['type' => 'date'],
+        'tags' => ['type' => 'string'],
+        'thread_id' => ['type' => 'long'],
+        'title' => ['type' => 'string'],
+        'title_unicode' => ['type' => 'string'],
+        'user_id' => ['type' => 'long'],
+        'video' => ['type' => 'boolean'],
+    ];
 
     /*
     |--------------------------------------------------------------------------
@@ -1097,142 +1154,5 @@ class Beatmapset extends Model
         }
 
         return Forum\Post::find($topic->topic_first_post_id);
-    }
-
-    public static function esIndexName()
-    {
-        return 'beatmaps';
-    }
-
-    public static function esType()
-    {
-        return 'beatmaps';
-    }
-
-    public function toEsJson()
-    {
-        return [
-            'index' => static::esIndexName(),
-            'type' => static::esType(),
-            'id' => $this->beatmapset_id,
-            'body' => $this->esJsonBody(),
-        ];
-    }
-
-    private function esJsonBody()
-    {
-        return array_merge(
-            $this->esBeatmapsetValues(),
-            ['difficulties' => $this->esBeatmapValues()]
-        );
-    }
-
-    private function esBeatmapsetValues()
-    {
-        $mappings = static::ES_MAPPINGS_BEATMAPSETS;
-
-        $values = [];
-        foreach ($mappings as $field => $mapping) {
-            $value = $this[$field];
-            if ($value instanceof Carbon) {
-                $value = $value->toIso8601String();
-            }
-
-            $values[$field] = $value;
-        }
-
-        return $values;
-    }
-
-    private function esBeatmapValues()
-    {
-        $mappings = static::ES_MAPPINGS_BEATMAPS;
-
-        $values = [];
-        // initialize everything to an array.
-        foreach ($mappings as $field => $mapping) {
-            $values[$field] = [];
-        }
-
-        foreach ($this->beatmaps as $beatmap) {
-            foreach ($mappings as $field => $mapping) {
-                $values[$field][] = $beatmap[$field];
-            }
-        }
-
-        return $values;
-    }
-
-    const ES_MAPPINGS_BEATMAPS = [
-        'approved' => ['type' => 'long'],
-        'beatmap_id' => ['type' => 'long'],
-        'countNormal' => ['type' => 'long'],
-        'countSlider' => ['type' => 'long'],
-        'countSpinner' => ['type' => 'long'],
-        'countTotal' => ['type' => 'long'],
-        'diff_approach' => ['type' => 'double'],
-        'diff_drain' => ['type' => 'double'],
-        'diff_overall' => ['type' => 'double'],
-        'diff_size' => ['type' => 'double'],
-        'difficultyrating' => ['type' => 'double'],
-        'hit_length' => ['type' => 'long'],
-        'passcount' => ['type' => 'long'],
-        'playcount' => ['type' => 'long'],
-        'playmode' => ['type' => 'long'],
-        'total_length' => ['type' => 'long'],
-        'version' => ['type' => 'string'],
-    ];
-
-    const ES_MAPPINGS_BEATMAPSETS = [
-        'approved' => ['type' => 'long'],
-        'approved_date' => ['type' => 'date'],
-        'artist' => ['type' => 'string'],
-        'artist_unicode' => ['type' => 'string'],
-        'bpm' => ['type' => 'double'],
-        'creator' => ['type' => 'string'],
-        'difficulty_names' => ['type' => 'string'],
-        'download_disabled' => ['type' => 'boolean'],
-        'epilepsy' => ['type' => 'boolean'],
-        'favourite_count' => ['type' => 'long'],
-        'filename' => ['type' => 'string'],
-        'filesize' => ['type' => 'long'],
-        'filesize_novideo' => ['type' => 'long'],
-        'genre_id' => ['type' => 'long'],
-        'language_id' => ['type' => 'long'],
-        'last_update' => ['type' => 'date'],
-        'offset' => ['type' => 'long'],
-        'play_count' => ['type' => 'long'],
-        'rating' => ['type' => 'double'],
-        'source' => ['type' => 'string'],
-        'star_priority' => ['type' => 'long'],
-        'storyboard' => ['type' => 'boolean'],
-        'submit_date' => ['type' => 'date'],
-        'tags' => ['type' => 'string'],
-        'thread_id' => ['type' => 'long'],
-        'title' => ['type' => 'string'],
-        'title_unicode' => ['type' => 'string'],
-        'user_id' => ['type' => 'long'],
-        'video' => ['type' => 'boolean'],
-    ];
-
-    public static function esMappings()
-    {
-        return array_merge(
-            static::ES_MAPPINGS_BEATMAPSETS,
-            ['difficulties' => ['properties' => static::ES_MAPPINGS_BEATMAPS]]
-        );
-    }
-
-    public static function esReindexAll($batchSize = 1000, $fromId = 0)
-    {
-        $startTime = time();
-
-        $baseQuery = static::withoutGlobalScopes()
-            ->with('beatmaps'); // note that the with query will run with the default scopes.
-
-        $count = static::esIndexEach($baseQuery, 'beatmapset_id', $batchSize, $fromId);
-
-        $duration = time() - $startTime;
-        \Log::info("Indexed {$count} records in {$duration} s.");
     }
 }
